refactor FairModel instantiation, ClimateParams, debug flag
